git,zsh: add desscription for git-clang-format/lfs/filter-repo/review and disable fallback to filename completion
Some parts of this completion's behaviour are configurable:

Say you got your own git sub-commands (git will run a program `git-foo'
when you run "git foo") and you want "git f<tab>" to complete that sub
commands name for you. You can make that sub-command known to the completion
via the user-command style:

    % zstyle ':completion:*:*:git:*' user-commands foo:'description for foo'

`user-commands' is a list style, so you can add any number of programs there.
The :description part is optional, so you could add all git-* programs from
your $path like this:

    % zstyle ':completion:*:*:git:*' user-commands ${${(M)${(k)commands}:#git-*}/git-/}

A better solution is to create a function _git-foo() to handle specific
completion for that command.  This also allows you to add command-specific
completion as well.  Place such a function inside an autoloaded #compdef file
and you should be all set.  You can add a description to such a function by
adding a line matching

    #description DESCRIPTION

as the second line in the file.  See
Completion/Debian/Command/_git-buildpackage in the Zsh sources for an
example.

When _git does not know a given sub-command (say `bar'), it falls back to
completing file names for all arguments to that sub command. I.e.:

    % git bar <tab>

...will complete file names. If you do *not* want that fallback to be used,
use the `use-fallback' style like this:

    % zstyle ':completion:*:*:git*:*' use-fallback false

[1]: https://github.com/zsh-users/zsh/blob/5fe498124d6469fadded7b9ad9eb64649b93f2de/Completion/Unix/Command/_git#L3-L38
[2]: https://llvm.org/docs/Contributing.html#how-to-submit-a-patch