git,zsh: add desscription for git-revise
Fixup Fast
```
$ git add .
$ git revise HEAD~~
```
Running `git revise $COMMIT` directly collects changes staged in the index, and directly applies them to the specified commit. Conflicts are resolved interactively, and a warning will be shown if the final state of the tree is different from what you started with!

With an extra `-e`, you can update the commit message at the same time, and `-a` will stage your changes, so you don't have to!
Split Commits
```
$ git revise -c $COMMIT
Select changes to be included in part [1]:
diff --git b/file.txt a/file.txt
<snip>

Apply this hunk to index [y,n,q,a,d,e,?]?
```
Sometimes, a commit needs to be split in two, perhaps because a change ended up in the wrong commit. The `--cut` flag (and cut interactive command) provides a fast way to split a commit in-place.

Running `git revise --cut $COMMIT` will start a `git add -p` style hunk selector, allowing you to pick changes for part 1, and the rest will end up in part 2.

No more tinkering around with edit during a rebase to split off that comment you accidentally added to the wrong commit!

[1]: https://mystor.github.io/git-revise.html